+++ sator/adapters/driven/repositories/vulnerability/nvd.py
@@ -2,7 +2,7 @@ from pathlib import Path
 from typing import Dict, Any, List
 
 from sator.core.models.vulnerability import Vulnerability
-from sator.core.ports.driven.vulnerability_repository import VulnerabilityRepositoryPort
+from sator.core.ports.driven.repositories.vulnerability import VulnerabilityRepositoryPort
 from sator.adapters.driven.repositories.vulnerability.mappers import NVDVulnerabilityMapper
 
 from nvdutils.loaders.json.default import JSONDefaultLoader
