+++ sator/adapters/driven/repositories/vulnerability/osv.py
@@ -1,7 +1,7 @@
 from typing import Dict, Any, List
 
 from sator.core.models.vulnerability import Vulnerability
-from sator.core.ports.driven.vulnerability_repository import VulnerabilityRepositoryPort
+from sator.core.ports.driven.repositories.vulnerability import VulnerabilityRepositoryPort
 from sator.adapters.driven.repositories.vulnerability.mappers import OSVVulnerabilityMapper
 
 from osvutils.core.loader import OSVDataLoader
