+++ sator/core/ports/driven/repositories/diff.py
