+++ sator/core/ports/driven/repositories/vulnerability.py
